moved lemonade data mocker to its own file
